% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{makeObservationsBinary}
\alias{makeObservationsBinary}
\title{Convert an OTU Count Table to Presence/Absence}
\usage{
makeObservationsBinary(physeq)
}
\arguments{
\item{physeq}{A phyloseq object.}
}
\value{
A phyloseq object.
}
\description{
Convert an OTU Count Table to Presence/Absence
}
