% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{makeObservationsRelative}
\alias{makeObservationsRelative}
\title{Convert an OTU Count Table to Relative Abundance}
\usage{
makeObservationsRelative(physeq)
}
\arguments{
\item{physeq}{A phyloseq object.}
}
\value{
A phyloseq object.
}
\description{
Convert an OTU Count Table to Relative Abundance
}
