% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{prevalenceFilter}
\alias{prevalenceFilter}
\title{Apply a prevalence filter}
\usage{
prevalenceFilter(physeq, threshold)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{threshold}{Prevalence threshold.}
}
\value{
A phyloseq object.
}
\description{
Reatin taxa observed in greater or equal to number of samples*threshold.
}
