% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{readQIIMEFolder}
\alias{readQIIMEFolder}
\title{Read QIIME 2 Output from a Folder}
\usage{
readQIIMEFolder(filepath)
}
\arguments{
\item{filepath}{A filepath}
}
\value{
A phyloseq object
}
\description{
Read QIIME 2 Output from a Folder
}
