% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{removeNonbacterial}
\alias{removeNonbacterial}
\title{Remove Contaminant Sequences from Bacterial Datasets}
\usage{
removeNonbacterial(physeq)
}
\arguments{
\item{physeq}{A phyloseq object.}
}
\value{
A phyloseq object.
}
\description{
Remove Contaminant Sequences from Bacterial Datasets
}
